+++ src/main/java/com/sms/springboot/repository/FacultyRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.sms.springboot.model.Faculty;
 
 @Repository
-public interface FacultyRepository extends JpaRepository<Faculty, Long> {
+public interface FacultyRepository extends JpaRepository<Faculty, Integer> {
 
 }
